requirements.txt
streamlit
langchain_groq 
langchain_community 
pypdf 
sentence_transformers
python-dotenv
chromadb
docx2txtLoader 
PyPDF2
python-docx
Docx2txt
